opcionMenu
actualizamos el Convert.ToInt16(Console.ReadLine()) a Convert.ToInt32(Console.ReadLine()) ya que nos generaba un error porque se "sobrecargaba"

+++ Programacion parcial 2/Program.cs
@@ -84,7 +84,7 @@ namespace programacionparcial2
                 Console.WriteLine("Escriba el número de la opción elegida del menú: ");
              
 
-                opcionMenu = Convert.ToInt16(Console.ReadLine());
+                opcionMenu = Convert.ToInt32(Console.ReadLine());
 
                 switch (opcionMenu)
                             {
